Run SQL migration
The SQL migration to add the `get_clients_with_project_count` function has been reviewed and is ready to be run. This function will be used to accurately count projects associated with each client, addressing the issue of incorrect "Connected Projects" counts on the clients page. The `ClientService.ts` file will be updated to utilize this new function.